+++ package.json
@@ -21,7 +21,7 @@
     "broccoli-sourcemap-concat": "^0.4.3",
     "broccoli-stew": "^0.2.1",
     "broccoli-uglify-js": "^0.1.3",
-    "ember-cli": "^0.2.0",
+    "ember-cli": "^1.13.8",
     "es6-module-transpiler-amd-formatter": "^0.3.0",
     "loader.js": "^3.2.0",
     "qunitjs": "^1.17.1"
